+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Disable most `backdrop-filter` properties, they caused too many FPS issues
 * Add the ability to hide widgets from the public
 * Allow overriding the language (`lang` attribute) of a widget
+* Fix config of container widgets not being saved
 
 ## 0.2.0 (2022-11-02)